pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nofomo"
version = "0.1.0"
description = "RSS daily digest CLI for Telegram"
requires-python = ">=3.12"
dependencies = [
  "beautifulsoup4>=4.12,<5",
  "feedparser>=6.0,<7",
  "PyYAML>=6.0,<7",
  "requests>=2.32,<3",
]

[project.optional-dependencies]
dev = ["pytest>=8.0,<9", "pytest-mock>=3.14,<4"]

[project.scripts]
nofomo = "nofomo.main:main"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
